owm website needs an API key
OWM seems to have updated the website so that the data can only be
accessed after creating an account and using an API key. I've created an
account for myself and have included the API key in the URL to make this
work, but this will need to be fixed for later. OWM does mention that it
can remove the need for API keys on a per-request basis if the
application is open source for example.