converting the per-website scripts to functions
updating the owm, voll, and yr scripts to be callable functions instead
of just scripts. This allows all of the code to be called at the same
time within one parent script - namely collect_data. This collect_data
script can then be placed in the scheduled cron-job to be run daily.

Currently the bbc script is not being called. The python code is being
run on the pythonanywhere servers which currently don't allow access to
the bbc weather servers. Conversation on-going.