updating scripts to work with sqlalchemy package
forecast_db_interface class is now updated to use the sqlalchemy package.
The current version continues to create and use the sqlite3 db, but this can be changed with the flip of a switch.
The code is therefore independent of db types.

The various weather scripts (BBC, OWM, YR, VOLL) are now updated to work with the new class