re-architecting the forecast_db_interface class
Previously the class would open a connection to the DB and was expected
to close the connection at the end. This was causing a lot of issues
with maintaining multiple connections since the python-anywhere only
allows 3 simultaneous connections. One connection was opened by the
flask app via its db vairable and another was opened by the web-app and
a third was opened by the forecas_db_interface class and debugging could
open more connections. To simplify this, the forecast_db_interface class
now doesn't open a connection to the MySQL DB, but rather imports the db
variable from flask_app and expects it to have a session variable. This
could probably be architected better, but for now, this works.

As a consequence, all the functions in the forecast_db_interface class
which maintained the DB are now redundant and so need to be removed. The
functions calling these class functions also need to be edited ... thus
the updates to the varios python scripts.