adding new function insertRow
new function insertRow in class forecast_db_interface. This adds a row
to the correct table and then commits the addition to the DB as well.
Therefore the various scripts don't need to import the db variable from
the flask_app but instead can rely on the forecast_db_interface entirely
for the DB interaction.